feat: working blogging system
uses frontmatter on an md file as format, that is parsed in the python
backend. Add new md files and it will render. It also computes the html
on the server so the frontend does not need to decode, making things
much faster client side.

I added a fastapi router, put it inside v1. Right now we have a content
router that serves the blog posts in the content/ folder.

Frontend has a decent looking blog page and blog view.