feat: added proper markdown rendering
To do this i had to stop rendering html on the server, it had to be
rendered on the client. That is the change made here. No longer
converting to html and sending html, sending just content which is the
markdown text, and the client figures out how to render it and style it
and all that jazz.